backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
CacheControl==0.14.2
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.4
django-allauth==65.4.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
firebase-admin==6.6.0
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
googlemaps==4.10.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
# jwt==1.3.1
kombu==5.4.2
msgpack==1.1.0
multidict==6.1.0
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyfcm==2.0.8
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.2.1
reportlab==4.4.1
requests==2.32.3
rsa==4.9
service-identity==24.2.0
setuptools==79.0.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
zope.interface==7.2


# aiohappyeyeballs==2.4.6
# aiohttp==3.11.13
# aiosignal==1.3.2
# amqp==5.3.1
# asgiref==3.8.1
# attrs==25.1.0
# autobahn==24.4.2
# Automat==25.4.16
# billiard==4.2.1
# CacheControl==0.14.2
# cachetools==5.5.2
# celery==5.4.0
# certifi==2025.1.31
# cffi==1.17.1
# channels==4.2.0
# channels_redis==4.2.1
# charset-normalizer==3.4.1
# click==8.1.8
# click-didyoumean==0.3.1
# click-plugins==1.1.1
# click-repl==0.3.0
# colorama==0.4.6
# constantly==23.10.4
# cryptography==44.0.2
# daphne==4.1.2
# Django==5.1.4
# django-allauth==65.4.1
# django-cors-headers==4.7.0
# djangorestframework==3.15.2
# djangorestframework_simplejwt==5.5.0
# firebase-admin==6.6.0
# frozenlist==1.5.0
# geographiclib==2.0
# geopy==2.4.1
# google-api-core==2.24.2
# google-api-python-client==2.163.0
# google-auth==2.38.0
# google-auth-httplib2==0.2.0
# google-cloud-core==2.4.3
# google-cloud-firestore==2.20.1
# google-cloud-storage==3.1.0
# google-crc32c==1.6.0
# google-resumable-media==2.7.2
# googleapis-common-protos==1.69.1
# googlemaps==4.10.0
# grpcio==1.71.0
# grpcio-status==1.71.0
# httplib2==0.22.0
# hyperlink==21.0.0
# idna==3.10
# incremental==24.7.2
# # jwt==1.3.1
# kombu==5.4.2
# msgpack==1.1.0
# multidict==6.1.0
# pillow==11.1.0
# prompt_toolkit==3.0.50
# propcache==0.3.0
# proto-plus==1.26.1
# protobuf==5.29.3
# psycopg2-binary==2.9.10
# pyasn1==0.6.1
# pyasn1_modules==0.4.1
# pycparser==2.22
# pyfcm==2.0.8
# PyJWT==2.9.0
# pyOpenSSL==25.0.0
# pyparsing==3.2.1
# python-dateutil==2.9.0.post0
# python-decouple==3.8
# redis==5.0.8
# requests==2.32.3
# rsa==4.9
# service-identity==24.2.0
# setuptools==79.0.0
# six==1.17.0
# sqlparse==0.5.3
# Twisted==24.11.0
# txaio==23.1.1
# typing_extensions==4.13.2
# tzdata==2024.2
# uritemplate==4.1.1
# urllib3==2.3.0
# vine==5.1.0
# wcwidth==0.2.13
# yarl==1.18.3
# zope.interface==7.2
